Fix racing condition in termination of Pool
This patch fixes a racing condition when using Pool.terminate()
My application is using imap() but aborting further processing after some iterations. Therefore I used terminate() to end the workers pool but the called join() of the result_handler was hanging sporadically.
I traced the issue down to the problem that somehow poll() signals new data in the outqueue but the preceding get() never returned due to ERROR_IO_PENDING in winapi of the queue hand endless hanging WaitForMultipleObjects().

Waiting at most for the two sentinels issued by the pool cleanup fixes the issue. Also the ````while sentinel_count < 2```` seems to me more  reasonable and plattform-independent than the old ````for i in range(10)````